# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_chat.ChannelDetailResponse:
    description: 채널 상세 응답 데이터
    properties:
      created_at:
        example: "2024-01-15T09:30:00Z"
        type: string
      description:
        example: 자유롭게 대화하는 채널입니다
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      member_count:
        example: 5
        type: integer
      members:
        items:
          $ref: '#/definitions/internal_chat.MemberBrief'
        type: array
      name:
        example: 일반 채팅방
        type: string
      owner:
        $ref: '#/definitions/internal_chat.UserBrief'
      owner_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      updated_at:
        example: "2024-01-15T10:00:00Z"
        type: string
    type: object
  internal_chat.ChannelListResponse:
    description: 채널 목록 응답 데이터
    properties:
      channels:
        items:
          $ref: '#/definitions/internal_chat.ChannelResponse'
        type: array
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
    type: object
  internal_chat.ChannelResponse:
    description: 채널 응답 데이터
    properties:
      created_at:
        example: "2024-01-15T09:30:00Z"
        type: string
      description:
        example: 자유롭게 대화하는 채널입니다
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      member_count:
        example: 5
        type: integer
      name:
        example: 일반 채팅방
        type: string
      owner:
        $ref: '#/definitions/internal_chat.UserBrief'
      owner_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      updated_at:
        example: "2024-01-15T10:00:00Z"
        type: string
    type: object
  internal_chat.ChatMessageResponse:
    description: 채팅 메시지 응답 데이터
    properties:
      channel_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      content:
        example: 안녕하세요!
        type: string
      created_at:
        example: "2024-01-15T09:30:00Z"
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      nickname:
        example: 홍길동
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
  internal_chat.CreateChannelRequest:
    description: 채널 생성 요청 데이터
    properties:
      description:
        example: 자유롭게 대화하는 채널입니다
        maxLength: 500
        type: string
      name:
        example: 일반 채팅방
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
  internal_chat.MemberBrief:
    description: 채널 멤버 정보
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      joined_at:
        example: "2024-01-15T09:30:00Z"
        type: string
      nickname:
        example: 홍길동
        type: string
    type: object
  internal_chat.MessageListResponse:
    description: 메시지 목록 응답 데이터
    properties:
      has_more:
        example: true
        type: boolean
      messages:
        items:
          $ref: '#/definitions/internal_chat.ChatMessageResponse'
        type: array
    type: object
  internal_chat.MessageResponse:
    description: 일반 메시지 응답
    properties:
      message:
        example: 작업이 완료되었습니다
        type: string
    type: object
  internal_chat.UpdateChannelRequest:
    description: 채널 수정 요청 데이터
    properties:
      description:
        example: 수정된 설명
        maxLength: 500
        type: string
      name:
        example: 수정된 채널명
        maxLength: 100
        minLength: 1
        type: string
    type: object
  internal_chat.UserBrief:
    description: 간략한 사용자 정보
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      nickname:
        example: 홍길동
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/jin-chillo/go-chat
  description: JWT 기반 인증과 WebSocket 실시간 채팅 기능을 제공하는 Go 백엔드 서버
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GoChat API
  version: "1.0"
paths:
...
  /channels:
    get:
      consumes:
      - application/json
      description: 전체 채널 목록을 페이지네이션하여 조회합니다
      parameters:
      - default: 1
        description: 페이지 번호
        in: query
        name: page
        type: integer
      - default: 20
        description: 페이지당 항목 수
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 채널 목록
          schema:
            $ref: '#/definitions/internal_chat.ChannelListResponse'
        "401":
          description: 인증 실패
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 서버 에러
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 채널 목록 조회
      tags:
      - Channels
    post:
      consumes:
      - application/json
      description: 새로운 채팅 채널을 생성합니다. 생성자가 자동으로 채널 소유자가 됩니다
      parameters:
      - description: 채널 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_chat.CreateChannelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 생성된 채널
          schema:
            additionalProperties:
              $ref: '#/definitions/internal_chat.ChannelResponse'
            type: object
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 서버 에러
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 채널 생성
      tags:
      - Channels
  /channels/{id}:
    delete:
      consumes:
      - application/json
      description: 채널을 삭제합니다. 채널 소유자만 삭제할 수 있습니다
      parameters:
      - description: 채널 ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 삭제 성공
          schema:
            $ref: '#/definitions/internal_chat.MessageResponse'
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 권한 없음
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 채널 없음
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 서버 에러
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 채널 삭제
      tags:
      - Channels
    get:
      consumes:
      - application/json
      description: 채널의 상세 정보와 멤버 목록을 조회합니다
      parameters:
      - description: 채널 ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 채널 상세 정보
          schema:
            additionalProperties:
              $ref: '#/definitions/internal_chat.ChannelDetailResponse'
            type: object
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 채널 없음
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 서버 에러
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 채널 상세 조회
      tags:
      - Channels
    put:
      consumes:
      - application/json
      description: 채널 정보를 수정합니다. 채널 소유자만 수정할 수 있습니다
      parameters:
      - description: 채널 ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: 수정할 채널 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_chat.UpdateChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 수정된 채널
          schema:
            additionalProperties:
              $ref: '#/definitions/internal_chat.ChannelResponse'
            type: object
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 권한 없음
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 채널 없음
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 서버 에러
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 채널 수정
      tags:
      - Channels
  /channels/{id}/join:
    post:
      consumes:
      - application/json
      description: 채널에 참여합니다
      parameters:
      - description: 채널 ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 참여 성공
          schema:
            $ref: '#/definitions/internal_chat.MessageResponse'
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 채널 없음
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 이미 멤버임
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 서버 에러
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 채널 참여
      tags:
      - Channels
  /channels/{id}/leave:
    post:
      consumes:
      - application/json
      description: 채널에서 퇴장합니다. 채널 소유자는 퇴장할 수 없습니다
      parameters:
      - description: 채널 ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 퇴장 성공
          schema:
            $ref: '#/definitions/internal_chat.MessageResponse'
        "400":
          description: 잘못된 요청 또는 멤버가 아님
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 소유자는 퇴장 불가
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 채널 없음
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 서버 에러
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 채널 퇴장
      tags:
      - Channels
  /channels/{id}/members/online:
    get:
      consumes:
      - application/json
      description: 채널에 현재 접속 중인 온라인 멤버 목록을 조회합니다
      parameters:
      - description: 채널 ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 온라인 사용자 목록
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 채널 없음
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 서버 에러
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 온라인 멤버 조회
      tags:
      - Channels
  /channels/{id}/messages:
    get:
      consumes:
      - application/json
      description: 채널의 메시지 이력을 커서 기반 페이지네이션으로 조회합니다
      parameters:
      - description: 채널 ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - default: 50
        description: 가져올 메시지 수 (최대 100)
        in: query
        name: limit
        type: integer
      - description: 이 시간 이전 메시지 조회 (RFC3339 형식)
        in: query
        name: before
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 메시지 목록
          schema:
            $ref: '#/definitions/internal_chat.MessageListResponse'
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 채널 없음
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 서버 에러
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 메시지 이력 조회
      tags:
      - Channels
  /channels/my:
    get:
      consumes:
      - application/json
      description: 현재 로그인한 사용자가 참여한 채널 목록을 조회합니다
      parameters:
      - default: 1
        description: 페이지 번호
        in: query
        name: page
        type: integer
      - default: 20
        description: 페이지당 항목 수
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 채널 목록
          schema:
            $ref: '#/definitions/internal_chat.ChannelListResponse'
        "401":
          description: 인증 실패
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 서버 에러
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 내 채널 목록 조회
      tags:
      - Channels
securityDefinitions:
  BearerAuth:
    description: 'Bearer 토큰을 입력하세요 (예: Bearer eyJhbGciOiJIUzI1NiIs...)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"